wip: move out convert_to_planner_data
