+++ backend/server/routers/ctf.py
@@ -46,7 +46,7 @@ from typing import Annotated, Callable, Optional
 
 from fastapi import APIRouter, Security
 from server.routers.utility.sessions.middleware import HTTPBearerToUserID
-from server.routers.utility.user import get_setup_user
+from server.routers.utility.user import convert_to_planner_data, get_setup_user
 from server.routers.model import ValidPlannerData
 
 router = APIRouter(
@@ -254,7 +254,6 @@ def validate_ctf(uid: Annotated[str, Security(require_uid)]):
     """
     Validates the CTF
     """
-    from server.routers.planner import convert_to_planner_data  # TODO: remove when converted
 
     data = convert_to_planner_data(get_setup_user(uid))
     passed: list[str] = []
